BASIC serial WIFI setup that is out of date
C64 BASIC PRG to setup the WIFI information and eventually the rest once I add it in. This PRG encodes the characters as needed and sends them over the serial port through UP9600 to setup the WIFI and other information on the ESP8266_Firmware. Planning on updating soon so I can build the binaries for release.